CONSOLE: Print frag messages to console, hide from notification area
If printflag PR_NONOTIFY is set, the line's time is set to 0 and so
is skipped when printing the notification area.  This means we have to
count backwards (skipping some lines) to find the one to start the notify
printing from.

Can turn on new handling of fragmessages with con_fragmessages 2